+++ src/WindowsUtils/ListModules.cpp
@@ -4,17 +4,18 @@
 
 #include "WindowsUtils/ListModules.h"
 
-#include <ObjectUtils/CoffFile.h>
-#include <OrbitBase/Logging.h>
-#include <OrbitBase/UniqueResource.h>
 #include <absl/base/casts.h>
 #include <windows.h>
+
+#include <filesystem>
+
+#include "ObjectUtils/CoffFile.h"
+#include "OrbitBase/Logging.h"
+#include "OrbitBase/UniqueResource.h"
 
 // Include after windows.h.
 #include <psapi.h>
 #include <tlhelp32.h>
-
-#include <filesystem>
 
 using orbit_object_utils::CoffFile;
