+++ src/WindowsUtils/include/WindowsUtils/ListProcesses.h
@@ -5,10 +5,10 @@
 #ifndef WINDOWS_UTILS_LIST_PROCESSES_H_
 #define WINDOWS_UTILS_LIST_PROCESSES_H_
 
-#include <OrbitBase/ThreadConstants.h>
-
 #include <string>
 #include <vector>
+
+#include "OrbitBase/ThreadConstants.h"
 
 namespace orbit_windows_utils {
 
@@ -16,10 +16,8 @@ struct Process {
   uint32_t pid = orbit_base::kInvalidProcessId;
   std::string name;
   std::string full_path;
-  std::string command_line;
   std::string build_id;
   bool is_64_bit = true;
-  double cpu_usage = 0;
 };
 
 // List all currently running processes.
